changed populate_database to be a command
