+++ jobs/management/commands/description.txt
@@ -5,9 +5,6 @@ Elitr oratio labitur sit ei, ei altera gloriatur interpretaris vel. Ea cum erant
 Percipit dignissim eu eos. Pri no modus disputando, eum no enim molestie. At lorem dolorem suscipit eos, ad alii eros semper quo. Labitur invidunt adipisci in est, est te omnes propriae intellegebat, commodo iudicabit sit at.
 Usu at diam populo, vel ei novum inermis. Ex duo eligendi instructior, eu ius accusam omittam. Est nisl tollit aliquam no, dolores fastidii apeirian mei cu. Duo ad modus noluisse tincidunt, dictas commodo luptatum sit et. Sea ei soluta indoctum, detracto placerat nominati ad nec. No natum ferri saepe sea, placerat urbanitas ea ius, simul laudem dissentiunt ad mei.
 Cu quo vivendum gloriatur. Qui ut modo purto rationibus. Sea cu senserit democritum reprehendunt, his stet movet maiestatis at. Te minimum vulputate contentiones has, ne vel nullam oblique similique. Eam nobis doming vivendo cu.
-
 Senserit cotidieque concludaturque his at, saepe dissentiunt est te, cu sit harum vidisse. Ornatus laoreet gubergren ius ex, duo numquam civibus id. Mel ei mutat numquam similique. An lorem utamur laboramus nec, his mollis cetero interesset no. Ut vix dico esse dolores, suscipit apeirian ea sit. Vis mazim verear no, ea solum tamquam nonumes usu. Duo dolorem percipitur eu, ne vim unum dissentiet vituperatoribus, clita eripuit an nec.
-
 Est saperet detraxit constituam cu. Odio mucius iisque ex quo. Ex vix putant suavitate, vidisse alienum euripidis mea cu, est in noster rationibus repudiandae. Duo platonem scriptorem ea.
-
 Deseruisse constituam ex sea, odio quas ad eum, scripta perpetua expetenda id quo. Id nam verear instructior disputationi, cu his sale tation ornatus. Sed affert possim voluptaria et, vel dissentias eloquentiam vituperatoribus ne, nec scripta minimum eu. Eum zril liberavisse te.
